backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
PyJWT==2.11.0
pymongo==4.7.2
python-dotenv==1.0.0
bcrypt==4.1.3
